ルーティング記述の簡略化と削除処理の条件追加

+++ app/Models/SevenColumn.php
@@ -273,6 +273,8 @@ class SevenColumn extends Model
     public function deleteSevencolumn($id)
     {
         $seven_column = SevenColumn::find($id);
-        $seven_column->delete();
+        if(Auth::id() === $seven_column->user_id) {
+            $seven_column->delete();
+        }
     }
 }
